# src/main/resources/api.yml
openapi: 3.0.1
info:
  title: Bank Account API
  description: API for managing bank accounts, including operations like creating, retrieving, depositing, withdrawing, and deleting bank accounts.
  version: 1.0.0
servers:
  - url: http://localhost:8080

paths:
  /accounts:
    post:
      summary: Create a new bank account
      description: Creates a new bank account for a customer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankAccount'
      responses:
        '201':
          description: Bank account created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccount'

    get:
      summary: Retrieve all bank accounts
      description: Retrieves a list of all bank accounts.
      responses:
        '200':
          description: A list of bank accounts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BankAccount'

  /accounts/{id}:
    get:
      summary: Retrieve a bank account by ID
      description: Retrieves details of a specific bank account by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Bank account found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccount'
        '404':
          description: Bank account not found.

    delete:
      summary: Delete a bank account by ID
      description: Deletes a specific bank account by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Bank account deleted successfully.
        '404':
          description: Bank account not found.

  /accounts/{accountId}/deposit:
    put:
      summary: Deposit an amount into a bank account
      description: Deposits a specified amount into a bank account.
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
        - name: amount
          in: query
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Deposit successful, balance updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccount'
        '400':
          description: Invalid deposit amount.

  /accounts/{accountId}/withdraw:
    put:
      summary: Withdraw an amount from a bank account
      description: Withdraws a specified amount from a bank account.
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
        - name: amount
          in: query
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Withdrawal successful, balance updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccount'
        '400':
          description: Invalid withdrawal amount.
        '409':
          description: Insufficient funds or balance below allowed limit.

components:
  schemas:
    BankAccount:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the bank account.
        accountNumber:
          type: string
          description: Unique account number.
        balance:
          type: number
          format: double
          description: Current balance of the bank account.
        accountType:
          type: string
          enum:
            - SAVINGS
            - CHECKING
          description: Type of bank account.
        customerId:
          type: integer
          description: The ID of the customer who owns the bank account.
      required:
        - accountType
        - balance
        - customerId